phase1_embedding/pyproject.toml
[project]
name = "vectordb-benchmark-phase1"
version = "1.0.0"
description = "Phase 1: 向量生成性能测试"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.12.0",
    "httpx>=0.27.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "h5py>=3.10.0",
    "pyarrow>=15.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "datasets>=2.16.0",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
    "pynvml>=11.5.0",
    "psutil>=5.9.0",
    "gpustat>=1.1.0",
    "plotly>=5.18.0",
    "jinja2>=3.1.3",
    "kaleido>=0.2.1",
    "pyyaml>=6.0.1",
    "colorlog>=6.8.0",
    "python-dateutil>=2.8.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["benchmarks", "cache", "data", "models"]

[dependency-groups]
dev = []
